admin_app/requirements_admin.txt
django
djangorestframework
psycopg2
python-dotenv